feat: add applyTheme utility
